movie-picker/requirements.txt
Pillow               
requests                    